requirements.txt
streamlit>=1.36
pandas>=2.2
numpy>=1.26
plotly>=5.22
openpyxl>=3.1
kaleido==0.2.1
reportlab>=4.0
Pillow>=10.4
pydantic>=2.7
openai>=1.53
requests>=2.32
pytest>=8.0
flake8>=7.0
